backend/src/gateway/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
isort==5.11.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg==3.1.7
psycopg-binary==3.1.6
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0

requests~=2.28.1